ipatool: Export commits in the order they're in the branch
GitHub API returns the commits in chronological order [1]. To ensure that the
patches will be applicable we need to order the as they are ordered in the
original branch.
In our use case each commit will have exacly one parent so we can use the SHA
values to order the commits properly.

[1] https://developer.github.com/v3/repos/commits/#response-3