makepki.py - add Authority Key Identifier
Together with SKI comes AKI which may not be necessary
for the moment in our script but would like to implement it just
for completeness sake and for potential future use.

Here we use from_issuer_public_key() method to get AKI for our root
CA and from_issuer_subject_key_identifier() method for subca.

https://tools.ietf.org/html/rfc5280#section-4.2.1.1

Signed-off-by: Michal Reznik <mreznik@redhat.com>