ipatool: explicitly pass backport branches
Merging the backport branches in backport() results in the pr-push
command attempting to "backport" the change to the PR's target
branch (which fails because patch has already been pushed).

Update backport() to take an explicit list of backport target
branches and update the call sites to set the branches based on the
relevant option only - `--backport` for `pr-push` and `--branch` for
`backport`.