Add support for known issues
In Pagure issues there is a field 'knownissue' ('true'/'false') that can
be used to specify whether this issue currently has no solution. A
changelog entry for the issue (either from commit message as a release
note or from the ticket 'changelog' field) can be used to document the
issue for release notes.

The note will be added to 'Known issues' section.