ipatool: use 3way merge for git am
Patching sometimes fails.  Nothing gets applied (no changes to
working tree) and no failed hunk indication is given.  I do not
understand the reason.  Perhaps it is a quirk triggered by passing
the patch on stdin instead of as a file.

Experimentally, I found that supplying the `--3way' option to git am
fixed the problem (the problem that occurred for me, at least).